+++ vllm/reasoning/__init__.py
@@ -5,14 +5,16 @@ from .abs_reasoning_parsers import ReasoningParser, ReasoningParserManager
 from .deepseek_r1_reasoning_parser import DeepSeekR1ReasoningParser
 from .granite_reasoning_parser import GraniteReasoningParser
 from .hunyuan_a13b_reasoning_parser import HunyuanA13BReasoningParser
-from .nemotron_reasoning_parser import NemotronReasoningParser
 from .qwen3_reasoning_parser import Qwen3ReasoningParser
+from .string_based_reasoning_parser import (NemotronReasoningParser,
+                                            get_string_based_reasoning_parser)
 
 __all__ = [
     "ReasoningParser",
     "ReasoningParserManager",
     "DeepSeekR1ReasoningParser",
     "GraniteReasoningParser",
+    "get_string_based_reasoning_parser",
     "HunyuanA13BReasoningParser",
     "NemotronReasoningParser",
     "Qwen3ReasoningParser",
